+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+**7.1.0**
+
+```
+IMP: Command 'caf': Sorted by field name
+IMP: Screen: Use less nodes
+IMP: Screen: Auto vacuum (No more than 750 lines)
+IMP: Command 'repeat': Try to be nice with the main thread
+IMP: Command 'fuzz': Support one2many fields and other improvements
+IMP: ParameterReader: Now can use slashes to avoid double quotes grouping (Example: "This \"is\" an example")
+IMP: Handle big results: split current query to not block/crash the browser
+IMP: Now can set default values for positional replacements (Example: $1[defaul text value] or $1[42])
+
+ADD: Command 'mute': Runs the command only printing errors. (Useful with the 'repeat' command)
+ADD: Command 'count': Get the number of records
+ADD: Command 'exportfile': Exports the command result to a text file
+ADD: New parameter generator: $FLOAT
+ADD: Prompt changes the color per host (locahost doesn't have any color)
+
+FIX: Changed some calls using the old 'print' call
+FIX: Command 'caf': Now can use 'fields' parameter
+FIX: Shadow Input: Correct sync. with input scroll left
+```
+
 **7.0.0**
 
 ```